+++ plugins/ibis/.ci_extend.sh
@@ -1,0 +1,1 @@
+python -m pip install plugins/sqlalchemy