+++ plugins/pillow/superduper_pillow/__init__.py
@@ -4,4 +4,6 @@ _, requirements = requires_packages(['PIL', '10.2.0', None, 'pillow'])
 
 from .encoder import pil_image
 
+__version__ = '0.3.0'
+
 __all__ = ['pil_image']