+++ superduper/__init__.py
@@ -8,7 +8,7 @@ ROOT = configs.ROOT
 
 logging = logger.Logging
 
-__version__ = '0.2.0'
+__version__ = '0.3.0'
 
 # from superduper.backends import ibis, mongodb