+++ superduper/backends/ibis/data_backend.py
@@ -1,1 +1,1 @@
-# from superduper_ibis.data_backend import *  # noqa
+from superduper_ibis.data_backend import *  # noqa