+++ superduper/backends/ibis/field_types.py
@@ -1,1 +1,1 @@
-# from superduper_ibis.field_types import *  # noqa
+from superduper_ibis.field_types import *  # noqa